refactor(jsruntime): use enum class

+++ libs/jsruntime/src/bridge/llvmir/compiler.cc
@@ -1592,7 +1592,7 @@ llvm::Value* Compiler::CreateIsNonNullish(const Item& item) {
 
 llvm::Value* Compiler::CreateIsNonNullish(llvm::Value* value_ptr) {
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  return builder_->CreateICmpUGT(kind, builder_->getInt8(ValueKind::Null));
+  return builder_->CreateICmpUGT(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Null)));
 }
 
 // 7.1.2 ToBoolean ( argument )
@@ -1729,12 +1729,12 @@ llvm::Value* Compiler::CreateIsStrictlyEqual(llvm::Value* x, llvm::Value* y) {
 
 llvm::Value* Compiler::CreateIsUndefined(llvm::Value* value_ptr) {
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  return builder_->CreateICmpEQ(kind, builder_->getInt8(ValueKind::Undefined));
+  return builder_->CreateICmpEQ(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Undefined)));
 }
 
 llvm::Value* Compiler::CreateIsNull(llvm::Value* value_ptr) {
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  return builder_->CreateICmpEQ(kind, builder_->getInt8(ValueKind::Null));
+  return builder_->CreateICmpEQ(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Null)));
 }
 
 llvm::Value* Compiler::CreateIsSameBooleanValue(llvm::Value* value_ptr, llvm::Value* value) {
@@ -1743,7 +1743,7 @@ llvm::Value* Compiler::CreateIsSameBooleanValue(llvm::Value* value_ptr, llvm::Va
   auto* merge_block = llvm::BasicBlock::Create(*context_, "", function_);
 
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(ValueKind::Boolean));
+  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Boolean)));
   builder_->CreateCondBr(cond, then_block, else_block);
 
   builder_->SetInsertPoint(then_block);
@@ -1769,7 +1769,7 @@ llvm::Value* Compiler::CreateIsSameNumberValue(llvm::Value* value_ptr, llvm::Val
   auto* merge_block = llvm::BasicBlock::Create(*context_, "", function_);
 
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(ValueKind::Number));
+  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Number)));
   builder_->CreateCondBr(cond, then_block, else_block);
 
   builder_->SetInsertPoint(then_block);
@@ -1795,7 +1795,7 @@ llvm::Value* Compiler::CreateIsSameFunctionValue(llvm::Value* value_ptr, llvm::V
   auto* merge_block = llvm::BasicBlock::Create(*context_, "", function_);
 
   auto* kind = CreateLoadValueKindFromValue(value_ptr);
-  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(ValueKind::Function));
+  auto* cond = builder_->CreateICmpEQ(kind, builder_->getInt8(static_cast<uint8_t>(ValueKind::Function)));
   builder_->CreateCondBr(cond, then_block, else_block);
 
   builder_->SetInsertPoint(then_block);
